refactor: extract shell action catalog
