+++ scripts/rust-source-size-baseline.txt
@@ -1,4 +1,4 @@
 # Rust source files above the 1,000-line target. Keep sorted by path.
 # A file may only stay equal to this maximum or shrink with this entry tightened.
-crates/shell-core/src/actions.rs 1591
+crates/shell-core/src/actions.rs 1240
 crates/shell-core/src/reducer.rs 2723
